# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Facebook Ads Marketing API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/password/create": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send Reset Password link to email",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password using reset link",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Reset Password Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/user/{userId}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update facebook token to long-lived token  after logged in to facebook, Fetch Ad accounts about by fb_user_id",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "fb_token",
                        "in": "query",
                        "description": "Facebook token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fb_user_id",
                        "in": "query",
                        "description": "Facebook User ID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "id": {
                                    "description": "UUID",
                                    "type": "integer"
                                },
                                "name": {
                                    "description": "UUID",
                                    "type": "integer"
                                },
                                "email": {
                                    "description": "UUID",
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}